# src/main/resources/static/swagger.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/stores/register":{"post":{"tags":["store-controller"],"operationId":"registerStore","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Store"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/register":{"post":{"tags":["auth-controller"],"operationId":"registerUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/register/admin":{"post":{"tags":["auth-controller"],"operationId":"registerAdmin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/login":{"post":{"tags":["auth-controller"],"operationId":"authenticateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/stores/nearby":{"get":{"tags":["store-controller"],"operationId":"getNearbyStores","parameters":[{"name":"latitude","in":"query","required":true,"schema":{"type":"number","format":"double"}},{"name":"longitude","in":"query","required":true,"schema":{"type":"number","format":"double"}},{"name":"radius","in":"query","required":true,"schema":{"type":"number","format":"double"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Store"}}}}}}}},"/api/stores/":{"get":{"tags":["store-controller"],"operationId":"getAllStores","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Store"}}}}}}}},"/api/":{"get":{"tags":["auth-controller"],"operationId":"home","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"Store":{"type":"object","properties":{"name":{"type":"string"},"address":{"type":"string"},"phone":{"type":"string"},"category":{"type":"string"},"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"}}},"User":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"role":{"type":"string"},"address":{"type":"string"},"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"}}}}}}